--- src/main/java/ladder/model/LadderGame.java
@@ -1,17 +1,0 @@
-package ladder.model;
-
-import java.util.*;
-
-public class LadderGame {
-
-    private final List<Player> players;
-    private final Height height;
-
-    public LadderGame(List<Player> players, Height height) {
-        this.players = players;
-        this.height = height;
-    }
-
-    public void play() {
-    }
-}
