+++ src/main/java/ladder/view/ResultView.java
@@ -1,10 +1,21 @@
 package ladder.view;
 
+import ladder.model.Line;
+import ladder.model.Player;
+
 import java.util.List;
 
 public class ResultView {
 
-    public static void printPlayerName(List<String> names) {
-        names.forEach(name -> System.out.printf("%5s ",name));
+    public static void printPlayerName(List<Player> players) {
+        players.forEach(player -> System.out.printf("%5s ", player.name()));
+        System.out.println();
+    }
+
+    public static void printLineList(List<Line> lineList) {
+        lineList.forEach((line) -> {
+            line.draw().forEach(System.out::print);
+            System.out.println();
+        });
     }
 }
